Move rigid group tests to md
